Add files via upload
More streamlined, effective version of the original essayrevii: simply just shuts off your computer. Less fun, though.
Text that appears before computer shuts off is a romanized version of "本当にバカだよな？", which translates to “You really are an idiot, aren’t you?”. This is a reference to the "You are an idiot" song, the virus that inspired essayrevii in the first place.